refactor(canvas-workspace): reorganize components into folder-per-component structure
Move all components from flat `components/*.tsx` to `components/${Name}/index.tsx`
and update cross-component relative imports accordingly.
