+++ apps/canvas-workspace/src/renderer/src/components/CanvasNodeView/index.tsx
@@ -1,9 +1,9 @@
 import { useCallback } from "react";
 import type { CanvasNode, FrameNodeData } from "../types";
 import type { ResizeEdge } from "../hooks/useNodeResize";
-import { FileNodeBody } from "./FileNodeBody";
-import { TerminalNodeBody } from "./TerminalNodeBody";
-import { FrameNodeBody, FrameColorPicker } from "./FrameNodeBody";
+import { FileNodeBody } from "../FileNodeBody";
+import { TerminalNodeBody } from "../TerminalNodeBody";
+import { FrameNodeBody, FrameColorPicker } from "../FrameNodeBody";
 
 interface Props {
   node: CanvasNode;
