+++ apps/canvas-workspace/src/renderer/src/components/FloatingToolbar/index.tsx
