+++ apps/canvas-workspace/src/renderer/src/components/FrameNodeBody/index.tsx
