+++ apps/canvas-workspace/src/renderer/src/components/SearchPalette/index.tsx
