+++ apps/canvas-workspace/src/renderer/src/components/Sidebar/index.tsx
