+++ apps/canvas-workspace/src/renderer/src/components/SlashCommandMenu/index.tsx
