+++ apps/canvas-workspace/src/renderer/src/components/TerminalNodeBody/index.tsx
