+++ apps/canvas-workspace/src/renderer/src/components/ZoomIndicator/index.tsx
